Fix double render bug while "updating" fume hood data
More specifically, when the main index page gets the fumehood data,
it now returns a dictionary with keys being an "id number" for
a hood and vals being their current energy consumption. When
index.html gets this response, it then uses jquery to find the
relevant fumehood energy widget to update the text value.

This fixes the double render since now it's not fetching the entire
html page again and rerendering it inside the paragraph. This still
fetches real time data, and better yet this method would only rerender
the parts of the html page that need updating instead of the entirety
of the page.

The old code is still there commented out, if you want to compare to
what was happening prior.